refactor: ui/ -> shadcn-ui/
