+++ components.json
@@ -13,7 +13,7 @@
   "aliases": {
     "components": "@/components",
     "utils": "@/lib/utils",
-    "ui": "@/components/ui",
+    "ui": "@/components/shadcn-ui",
     "lib": "@/lib",
     "hooks": "@/hooks"
   },
