+++ src/app/not-found.tsx
@@ -3,7 +3,7 @@
 import { ArrowLeft, Home } from 'lucide-react';
 import Link from 'next/link';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn-ui/button';
 
 export default function NotFound() {
   return (
