+++ src/components/layout/mode-switch.tsx
@@ -4,7 +4,7 @@ import { useTheme } from 'next-themes';
 import { useEffect, useState } from 'react';
 
 import { Icons } from '@/components/icons';
-import { Switch } from '@/components/ui/switch';
+import { Switch } from '@/components/shadcn-ui/switch';
 
 export function ModeSwitch() {
   const { resolvedTheme, setTheme } = useTheme();
