+++ src/components/layout/site-footer.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 
 import { Icons, SocialIcons } from '@/components/icons';
-import { buttonVariants } from '@/components/ui/button';
+import { buttonVariants } from '@/components/shadcn-ui/button';
 import { siteConfig } from '@/config/site';
 import { cn } from '@/lib/utils';
